requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-rosetta==0.10.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
polib==1.2.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
timeago==1.0.16
